Adding mysqli_connect.inc.php file
This should let us connect to the database.  I haven't tested it though,
so let me know if there are any problems when someone tries to use it.
Also, at some point, I think we may need to change what it logs in
under, but that'll be something to talk to Bruce about.